Add files via upload
This program or a mini-game is designed when you don't have anyone to play with or you are under lockdown alone. There are a number of functions that this program requires so let us have an overview of each.

• a random function: to generate rock, paper, or

scissors. • valid function: to check the validity of the move. • result function: to declare the winner of the round.

• scorekeeper: to keep track of the score.

The program requires the user to make the first move before it makes one the move. Once the move is validated the input is evaluated, the input entered could be a string or an alphabet. After evaluating the input string a winner is decided by the result function and the score of the round is updated by the scorekeeper function.